requirements.txt
pillow==4.3.0